polish: add dp kind specification to real declarations (fixes #2581) (#2602)
* polish: add dp kind specification to real declarations (fixes #2581)

Add explicit real(dp) type specification to 18 real declarations
across memory, performance, semantic, and AST subsystems per
CLAUDE.md Fortran rules:

- src/memory/arena_memory.f90: 1 type field + 1 literal
- src/memory/compiler_arena.f90: 3 type fields + 7 literals
- src/performance/ast_performance.f90: 5 type fields + 2 literals
- src/semantic/types/type_system_arena.f90: 1 type field
- src/semantic/convergence_checker.f90: 2 type fields
- src/semantic/analyzers/base_analyzer.f90: 1 type field
- src/ast/arena/ast_arena_core.f90: 4 type fields
- src/ast/ast_base.f90: 1 type field
- src/ast/ast_node_registry.f90: 1 type field

This ensures numerical precision consistency by using double
precision (real64) for all floating-point metrics and statistics.

* fix: use real(dp) for tolerance in results_equals

The confidence_score field was updated to real(dp) but the local
tolerance parameter in results_equals was still using plain real.
This ensures consistent precision for the comparison.

+++ src/ast/arena/ast_arena_core.f90
@@ -3,7 +3,7 @@ module ast_arena_core
     ! Focused on essential arena operations without compatibility layer
 
     use ast_base, only: ast_node
-    use, intrinsic :: iso_fortran_env, only: int64
+    use, intrinsic :: iso_fortran_env, only: int64, dp => real64
     use arena_memory, only: base_arena_t, arena_handle_t, arena_checkpoint_t
     implicit none
     private
@@ -33,7 +33,7 @@ module ast_arena_core
         character(len=256) :: string_data = ""  ! String content (names, literals)
         integer :: integer_data = 0  ! Integer content (indices, counts)
         logical :: boolean_data = .false.  ! Boolean flags
-        real :: real_data = 0.0  ! Real values
+        real(dp) :: real_data = 0.0_dp  ! Real values
 
         ! Extended metadata
         integer :: source_line = 0  ! Source code line number
@@ -104,15 +104,15 @@ module ast_arena_core
     type :: ast_arena_stats_t
         integer :: node_count = 0  ! Number of AST nodes
         integer(int64) :: total_memory = 0  ! Total memory usage
-        real :: utilization = 0.0  ! Memory utilization ratio
+        real(dp) :: utilization = 0.0_dp  ! Memory utilization ratio
         integer :: total_allocations = 0  ! Allocation counter
         integer :: total_validations = 0  ! Validation counter
-        real :: allocation_rate = 0.0  ! Nodes per second
+        real(dp) :: allocation_rate = 0.0_dp  ! Nodes per second
         integer :: max_depth = 0  ! Maximum tree depth
         integer :: total_children = 0  ! Total child relationships
         integer :: active_nodes = 0  ! Number of active (non-freed) nodes
         integer :: freed_nodes = 0  ! Number of freed nodes
-        real :: fragmentation = 0.0  ! Memory fragmentation ratio
+        real(dp) :: fragmentation = 0.0_dp  ! Memory fragmentation ratio
 
         ! Compatibility fields for old arena API
         integer :: total_nodes = 0  ! Alias for node_count
